# Fix broken cache eviction in clockcache