@niksajakovljevic 2022-08-31
Just wondering if you managed to benchmark version of code that does not split array?