@Harkishen-Singh 2022-09-01
Is https://github.com/timescale/promscale/pull/1603#issuecomment-1232919086 benchmark before the fix or after the fix. If its after the fix, then we are loosing perf in `Eviction`, right?